requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
anthropic==0.7.7
